pyproject.toml
[project]
name = "aicoursesummaraizer"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "3.11.*"
dependencies = [
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "openai (>=1.97.1,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"  # Match your Python version
plugins = ['pydantic.mypy']
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = false
strict_optional = true
no_implicit_optional = true
show_error_codes = true
exclude = "tests/*|.venv/*|__pycache__/*"  # Exclude test/venv files


[tool.pylint]
load-plugins = "pylint_pydantic"
